+++ mobile/src/environments/environment.prod.ts
@@ -2,6 +2,6 @@ export const environment = {
   production: true,
   apiUrl: 'http://192.168.1.72:8081',
   appName: 'Elykia',
-  version: '2.28.17'
+  version: '2.29.1'
 };